Pre-Tournament
On your mark...

As a side note, further git pushes will not reflect every successive
change. We are using ftp to manage the prouduction site. Git is used as
a decentralized backup. (to take control, first pull/sync your local
files with whatever is on github, then use ftp to change the godaddy
server.